Initial work: starting mobile first, moved logic into mobile menu
