+++ app/components/ui/SiteSearchInput.module.scss
@@ -15,7 +15,7 @@ $focus-height: calc(#{$input-height} - (#{$border-width} * 2));
     border-right: none;
     border-radius: $rounded-full 0 0 $rounded-full;
     flex: 1;
-    width: 500px;
+    width: 400px;
     max-width: 100%;
     height: $input-height;
